mod:: 로그인, 회원가입 컨트롤러, 서비스 계층명 변경
- 로그인 회원가입 컨트롤러, 서비스 계층명 변경(User -> UserAuth)
- 성실도 점수 API를 위한 UserController, UserService 파일 생성
