+++ ontime-back/src/main/java/devkor/ontime_back/controller/UserController.java
@@ -1,72 +1,29 @@
 package devkor.ontime_back.controller;
 
-import devkor.ontime_back.dto.UserAdditionalInfoDto;
-import devkor.ontime_back.dto.UserSignUpDto;
+import devkor.ontime_back.service.ScheduleService;
 import devkor.ontime_back.service.UserService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.Map;
-
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/user")
 @RequiredArgsConstructor
 public class UserController {
 
     private final UserService userService;
+    private final ScheduleService scheduleService;
 
-    @Operation(summary = "회원가입")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "회원가입 성공", content = @Content(schema = @Schema(implementation = String.class))),
-            @ApiResponse(responseCode = "400", description = "잘못된 요청", content = @Content(schema = @Schema(implementation = String.class)))
-    })
-    @PostMapping("/sign-up")
-    public String signUp(@RequestBody UserSignUpDto userSignUpDto) throws Exception {
-        userService.signUp(userSignUpDto);
-        return "회원가입 성공";
+    // 성실도 페이지 데이터 반환
+    @GetMapping("/{userId}/punctuality")
+    public PunctualityPageResponse getPunctualityPage(@PathVariable Long userId) {
+        float punctualityScore = userService.getPunctualityScore(userId);
+        List<LatenessHistoryResponse> latenessHistory = scheduleService.getLatenessHistory(userId);
+        List<ScheduleHistoryResponse> scheduleHistory = scheduleService.getScheduleHistory(userId);
+
+        return new PunctualityPageResponse(punctualityScore, latenessHistory, scheduleHistory);
     }
-
-
-    @Operation(summary = "User Login", description = "로그인 요청을 통해 JWT 토큰을 발급받습니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "로그인 성공", content = @Content(schema = @Schema(implementation = String.class))),
-            @ApiResponse(responseCode = "401", description = "인증 실패", content = @Content(schema = @Schema(implementation = String.class)))
-    })
-    @PostMapping("/login")
-    public String login(
-            @io.swagger.v3.oas.annotations.parameters.RequestBody(
-                    description = "로그인 요청 JSON 데이터",
-                    required = true,
-                    content = @Content(
-                            schema = @Schema(type = "object", example = "{\"email\": \"user@example.com\", \"password\": \"password123\"}")
-                    )
-            )
-            @RequestBody Map<String, String> loginRequest
-    ) {
-        return "로그인 요청 성공"; // 실제 로그인 처리는 Security 필터에서 수행
-    }
-
-
-    @Operation(summary = "JWT 테스트")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "JWT 테스트 성공", content = @Content(schema = @Schema(implementation = String.class)))
-    })
-    @GetMapping("/jwt-test")
-    public String jwtTest() {
-        return "jwtTest 요청 성공";
-    }
-
-    @PutMapping("/{id}/additional-info")
-    public ResponseEntity<?> addInfo(@PathVariable Long id, @RequestBody UserAdditionalInfoDto userAdditionalInfoDto) throws Exception {
-        userService.addInfo(id, userAdditionalInfoDto);
-        return ResponseEntity.ok("추가 정보 기입 성공");
-    }
-
 }
+
