+++ ontime-back/src/main/java/devkor/ontime_back/entity/User.java
@@ -3,8 +3,6 @@ package devkor.ontime_back.entity;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import java.sql.Time;
-import java.util.UUID;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
